+++ frontend/luximia_erp_ui/app/(configuraciones)/ajustes/page.jsx
@@ -1,8 +1,8 @@
 'use client';
 import { useState, useEffect } from 'react';
 import Image from 'next/image';
-import { useAuth } from '../../../context/AuthContext.jsx';
-import { getUser, updateUser } from '../../../services/api.js';
+import { useAuth } from '@/context/AuthContext';
+import { getUser, updateUser } from '@/services/api';
 
 export default function AjustesPage() {
     const { user } = useAuth();
