+++ frontend/luximia_erp_ui/app/(configuraciones)/auditoria/page.jsx
@@ -1,8 +1,8 @@
 'use client';
 import React, { useEffect, useState } from 'react';
-import ReusableTable from '../../components/ReusableTable';
-import { useAuth } from '../../context/AuthContext.jsx';
-import { getAuditLogs, downloadAuditLogExcel } from '../../services/api';
+import ReusableTable from '@/components/ui/tables/ReusableTable';
+import { useAuth } from '@/context/AuthContext';
+import { getAuditLogs, downloadAuditLogExcel } from '@/services/api';
 
 const COLUMNAS = [
   { header: 'Usuario', render: row => row.user || '-' },
