+++ frontend/luximia_erp_ui/app/(operaciones)/contratos/page.jsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 import Link from 'next/link';
-import { getContratos } from '../../../services/api';
+import { getContratos } from '@/services/api';
 
 export default function ContratosPage() {
     const [contratos, setContratos] = useState([]);
